+++ client/lib/analytics.js
@@ -56,9 +56,27 @@ var analyticsInit = function() {
     })();
   }
 
+  // Google Analytics
+  if ((googleAnalyticsId = getSetting("googleAnalyticsId"))){
+    window._gaq = window._gaq || [];
+    _gaq.push(['_setAccount', googleAnalyticsId]);
+
+    (function() {
+      var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
+      ga.src = ('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';
+      var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
+    })();
+  } 
+
 }
 
 var analyticsRequest = function() {
+
+  // Google Analytics
+  if ((googleAnalyticsId = getSetting("googleAnalyticsId"))){
+    window._gaq.push(['_trackPageview', window.location.pathname]);
+  } 
+
   // Mixpanel
   if(typeof mixpanel !== 'undefined' && typeof mixpanel.people !== 'undefined'){
     if(Meteor.user()){